test community api endpoint without hitting database
