--- server/__mockData__/verifiedUser.json
@@ -1,63 +1,0 @@
-{
-  "user": {
-    "fccCerts": {
-      "Back_End": false,
-      "Data_Visualization": false,
-      "Front_End": true
-    },
-    "career": {
-      "hasBeenEmployed": "",
-      "jobSearch": "",
-      "tenure": "1-3 years",
-      "company": "Facebook, Google",
-      "working": "yes"
-    },
-    "social": {
-      "twitter": "",
-      "linkedin": "Peter Weinberg",
-      "codepen": "no_stack_dub_sack"
-    },
-    "projects": [
-      {
-        "label": "https://github.com/",
-        "item": "FCC-Alumni/alumni-network"
-      }
-    ],
-    "skillsAndInterests": {
-      "codingInterests": [
-        "AWS",
-        "Bitcoin / Digital Currency",
-        "Behavior Driven Development",
-        "Backend Development"
-      ],
-      "coreSkills": [
-        "Angular",
-        "C#",
-        "Bash",
-        "C++"
-      ]
-    },
-    "mentorship": {
-      "mentorshipSkills": "writing code",
-      "isMentee": true,
-      "isMentor": true
-    },
-    "personal": {
-      "flag": "us",
-      "country": "United States",
-      "bio": "a person that lives in the world",
-      "location": "Greater NYC Area",
-      "email": {
-        "email": "pweinberg555@coolmail.com",
-        "private": true
-      },
-      "displayName": "Peter Weinberg",
-      "profileUrl": "https://github.com/no-stack-dub-sack",
-      "avatarUrl": "https://avatars2.githubusercontent.com/u/18563015?v=4",
-      "memberSince": "2017-07-19T02:36:02.912Z"
-    },
-    "verifiedUser": true,
-    "username": "no-stack-dub-sack",
-    "githubId": "18563015"
-  }
-}
