+++ server/helpers/checkHonoraryMemberList.js
@@ -4,7 +4,7 @@ export default (username) => {
   return new Promise(resolve => {
     HonoraryMember.findOne({ username: username.toLowerCase() },
     (err, user) => {
-      if (err) {throw err;}
+      if (err) throw err;
       if (user) {
         resolve(true);
       } else {
