+++ server/helpers/checkWhiteList.js
@@ -4,7 +4,7 @@ export default (githubUsername) => {
   return new Promise(resolve => {
     WhiteListedUser.findOne({ githubUsername },
       (err, user) => {
-      if (err) {throw err;}
+      if (err) throw err;
       if (user) {
         resolve(user.fccUsername);
       } else {
