+++ server/helpers/handleProcessedUser.js
@@ -4,7 +4,7 @@ export default (certs, mongoId, req, res, username) => {
   if (!certs) {
     // user not verified, res with error
     User.findById(mongoId, (err, user) => {
-      if (err) {throw err;}
+      if (err) throw err;
       user.verifiedUser = false;
       user.save();
       res.status(401).json({ error: 'User cannot be verified' });
@@ -12,7 +12,7 @@ export default (certs, mongoId, req, res, username) => {
   } else {
     // verified user, proceed
     User.findById(mongoId, (err, user) => {
-      if (err) {throw err;}
+      if (err) throw err;
       /* we need to overwrite their session username too
       (only matters for whitelisted users) */
       req.user.username = username;
