+++ server/helpers/safeHandler.js
@@ -1,4 +1,3 @@
-
 /* utilize this function as a passthrough for async express
 callbacks to catch any unhandled errors so server won't hang */
 export default (cb) => {
